feat: add UpdateRound method to Mutation class
A new method `UpdateRound` has been implemented in the
`Mutation` class within `Mutation.cs`. This method allows
for updating the next round of matches in a bracket by
checking if all matches in the current round have been
played. If they have, it creates new matches for the
following round based on the winners and saves the
changes to the database, returning the updated bracket.
This enhances the functionality of the bracket management
system.